Added grdual recursive regularization, WARNING: the implementaion is hackish. Improved logging.
It is not clear if this recursive regularization helps. The idea is to get
a rough (and fast) result using strong regularization. But in practive the
optimization gets stuck on the rough (smooth) result. This can be overcome
by using strong stop criterions for the optimization, but then it converges
to noisy results and it is not clear if we gain anything. Not through comparision.